move attrs under shapes
