+++ src/aioway/_media/images.py
@@ -11,7 +11,7 @@ from torchvision import io as vio
 from torchvision.transforms import v2 as tt
 
 from aioway._utils import current_fake_mode, torch_set_fake_mode_func
-from aioway.attrs import Attr, AttrLike
+from aioway.spaces import Attr, AttrLike
 from aioway.dsets import IsImageTag
 
 from ._bases import TorchCompatible
