+++ src/aioway/dsets/tags/media.py
@@ -8,7 +8,7 @@ from collections import abc as cabc
 import torch
 
 from aioway._utils import is_fake_tensor
-from aioway.attrs import Attr, DType
+from aioway.spaces import Attr, DType
 
 from .tags import TensorTag, tags_dcls
