+++ src/aioway/spaces/README.md
@@ -1,0 +1,3 @@
+# `Space`s
+
+Spaces are essentially contracts to how a `torch.Tensor` should behave.
