+++ src/aioway/spaces/attrs/_bases.py
