+++ src/aioway/spaces/attrs/attrs.py
