+++ src/aioway/spaces/attrs/devices.py
