+++ src/aioway/spaces/attrs/dtypes.py
