+++ src/aioway/spaces/attrs/shapes.py
