+++ tests/dsets/tags/test_attrs.py
@@ -3,7 +3,7 @@
 import pytest
 import torch
 
-from aioway.attrs import Attr
+from aioway.spaces import Attr
 from aioway.dsets import AttrTag, TagDict
